Refactor welcome screen redirection logic
