+++ frontend/apps/console/src/layouts/DashboardLayout.tsx
@@ -46,10 +46,9 @@ import {
   UsersRound,
   Workflow,
 } from '@wso2/oxygen-ui-icons-react';
-import {useEffect, useMemo, useState, type ReactNode} from 'react';
+import {useMemo, type ReactNode} from 'react';
 import {useTranslation} from 'react-i18next';
 import {Link as NavigateLink, Outlet, useNavigate} from 'react-router';
-import getWelcomeDismissedStorageKey from '../features/welcome/utils/getWelcomeDismissedStorageKey';
 
 function ExportConfigButton(): ReactNode {
   const {t} = useTranslation();
@@ -81,23 +80,10 @@ function ExportConfigButton(): ReactNode {
 
 export default function DashboardLayout(): ReactNode {
   const {signIn, clearSession, discovery} = useAsgardeo();
-  const {config, isTrustedIssuerGenericOidc, getTrustedIssuerClientId, getClientUrl} = useConfig();
+  const {isTrustedIssuerGenericOidc, getTrustedIssuerClientId, getClientUrl} = useConfig();
   const {t} = useTranslation();
   const logger = useLogger();
   const navigate = useNavigate();
-  const [welcomeCheckComplete, setWelcomeCheckComplete] = useState(false);
-
-  useEffect(() => {
-    const productName = config.brand.product_name;
-    const dismissed = sessionStorage.getItem(getWelcomeDismissedStorageKey(productName)) === 'true';
-
-    if (!dismissed) {
-      sessionStorage.setItem(getWelcomeDismissedStorageKey(productName), 'true');
-      void navigate('/welcome', {replace: true});
-    } else {
-      setWelcomeCheckComplete(true);
-    }
-  }, [navigate, config.brand.product_name]);
 
   const handleSignOut = (signOut: () => Promise<void>): void => {
     if (isTrustedIssuerGenericOidc()) {
@@ -298,7 +284,9 @@ export default function DashboardLayout(): ReactNode {
         </Sidebar>
       </AppShell.Sidebar>
 
-      <AppShell.Main>{welcomeCheckComplete ? <Outlet /> : null}</AppShell.Main>
+      <AppShell.Main>
+        <Outlet />
+      </AppShell.Main>
 
       <AppShell.Footer>
         <Footer>
